safari: TabFSServer subprocess that can live long. fixes bug where fs would die after a minute or two
